caml-mimic/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_pytorch_select=0.1=cpu_0
aiohttp=3.7.2=py37h9ed2024_1
appnope=0.1.0=py37_0
argon2-cffi=20.1.0=py37haf1e3a3_1
async-timeout=3.0.1=py37_0
async_generator=1.10=py37h28b3542_0
attrs=20.3.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.2.1=py_0
boto=2.49.0=py37_0
boto3=1.9.66=py37_0
botocore=1.12.67=py37_0
bottleneck=1.3.4=py37h67323c0_0
bzip2=1.0.8=h1de35cc_0
ca-certificates=2022.3.29=hecd8cb5_1
cachetools=4.1.1=py_0
certifi=2021.10.8=py37hecd8cb5_2
cffi=1.14.3=py37h2125817_2
chardet=3.0.4=py37hecd8cb5_1003
click=7.1.2=py_0
cryptography=3.2.1=py37hbcfaee0_1
cycler=0.10.0=py37_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
docutils=0.16=py37_1
entrypoints=0.3=py37_0
ffmpeg=4.3=h0a44026_0
freetype=2.10.4=ha233b18_0
gensim=3.8.0=py37h6440ff4_0
gettext=0.21.0=h7535e17_0
gmp=6.2.1=h23ab428_2
gnutls=3.6.15=hed9c0bf_0
google-api-core=1.22.2=py37_0
google-auth=1.23.0=pyhd3eb1b0_0
google-cloud-core=1.4.3=pyhd3eb1b0_1
google-cloud-storage=1.33.0=pyhd3eb1b0_0
google-crc32c=1.0.0=py37haf1e3a3_0
google-resumable-media=1.1.0=py_1
googleapis-common-protos=1.52.0=py37_0
icu=58.2=h0a44026_3
idna=2.10=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2019.4=233
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=5.8.0=py37_1
ipython_genutils=0.2.0=py37_0
jinja2=2.11.2=py_0
jmespath=0.10.0=py_0
joblib=0.17.0=py_0
jpeg=9b=he5867d9_2
json5=0.9.5=py_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.0=py37hecd8cb5_0
jupyterlab=2.2.9=py_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=1.2.0=py_0
kiwisolver=1.3.0=py37h23ab428_0
lame=3.100=h1de35cc_0
lcms2=2.11=h92f6f08_0
libcrc32c=1.1.1=hb1e8313_2
libcxx=12.0.0=h2f01273_0
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libidn2=2.3.2=h9ed2024_0
libmklml=2019.0.5=0
libpng=1.6.37=ha441bb4_0
libprotobuf=3.13.0.1=hab81aa3_0
libsodium=1.0.18=h1de35cc_0
libtasn1=4.16.0=h9ed2024_0
libtiff=4.1.0=hcb84e12_1
libunistring=0.9.10=h9ed2024_0
libuv=1.40.0=haf1e3a3_0
libxml2=2.9.12=hcdb78fc_0
llvm-openmp=12.0.0=h0dcd299_1
lz4-c=1.9.2=h79c402e_3
markupsafe=1.1.1=py37h1de35cc_0
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py37h181983e_0
mistune=0.8.4=py37h1de35cc_0
mkl=2019.4=233
mkl-service=2.3.0=py37hfbe908c_0
mkl_fft=1.2.0=py37hc64f4ea_0
mkl_random=1.1.1=py37h959d312_0
multidict=4.7.6=py37haf1e3a3_1
nbclient=0.5.1=py_0
nbconvert=6.0.7=py37_0
nbformat=5.0.8=py_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.4.3=pyhd3eb1b0_0
nettle=3.7.3=h230ac6f_1
ninja=1.10.2=hf7b0b51_1
nltk=3.5=py_0
notebook=6.1.4=py37_0
numexpr=2.7.3=py37h16bde0e_0
numpy=1.19.2=py37h456fd55_0
numpy-base=1.19.2=py37hcfb5961_0
olefile=0.46=py37_0
openh264=2.1.0=hd9629dc_0
openssl=1.1.1n=hca72f7f_0
packaging=20.4=py_0
pandas=1.3.4=py37h743cdd8_0
pandoc=2.11=h0dc7051_0
pandocfilters=1.4.3=py37hecd8cb5_1
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=py37_0
pillow=8.0.1=py37h5270095_0
pip=20.2.4=py37hecd8cb5_0
pip-autoremove=0.9.1=pypi_0
prometheus_client=0.8.0=py_0
prompt_toolkit=1.0.15=py_1
protobuf=3.13.0.1=py37hb1e8313_1
ptyprocess=0.6.0=pyhd3eb1b0_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pygments=2.7.2=pyhd3eb1b0_0
pyopenssl=19.1.0=pyhd3eb1b0_1
pyparsing=2.4.7=py_0
pyrsistent=0.17.3=py37haf1e3a3_0
pysocks=1.7.1=py37hecd8cb5_0
python=3.7.9=h26836e1_0
python-dateutil=2.8.1=py_0
pytorch=1.9.1=py3.7_0
pytz=2020.1=py_0
pyzmq=19.0.2=py37hb1e8313_1
readline=8.0=h1de35cc_0
regex=2020.10.15=py37haf1e3a3_0
requests=2.24.0=py_0
rsa=4.6=py_0
s3transfer=0.1.13=py37_0
scikit-learn=0.24.2=py37hb2f4e1b_0
scipy=1.5.2=py37h912ce22_0
seaborn=0.11.2=pyhd3eb1b0_0
send2trash=1.5.0=py37_0
setuptools=50.3.1=py37hecd8cb5_1
simplegeneric=0.8.1=py37_2
six=1.15.0=py37hecd8cb5_0
smart_open=3.0.0=py_0
sqlite=3.33.0=hffcf06c_0
terminado=0.9.1=py37_0
testpath=0.4.4=py_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.10=hb0a8c7a_0
torchaudio=0.9.1=py37
torchtext=0.10.1=py37
torchvision=0.10.1=py37_cpu
tornado=6.0.4=py37h1de35cc_1
tqdm=4.51.0=pyhd3eb1b0_0
traitlets=5.0.5=py_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
urllib3=1.24.3=py37_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
wheel=0.35.1=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
yarl=1.5.1=py37haf1e3a3_0
zeromq=4.3.3=hb1e8313_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
